Crypto.Cipher.Camellia

Description

this only cover Camellia 128 bits for now, API will change once 192 and 256 mode are implemented too

Documentation

data Key

Constructors

Key

Fields

k :: Vector Word64

kw :: Vector Word64

ke :: Vector Word64

initKey128 :: ByteString -> Either String Key

encrypt :: Key -> ByteString -> ByteString

encrypt with the key a bytestring and returns the encrypted bytestring

decrypt :: Key -> ByteString -> ByteString

decrypt with the key a bytestring and returns the encrypted bytestring